[FEAT] #148 - delete unuseful code
