+++ app/src/main/java/com/playtogether_android/app/presentation/ui/onboarding/SearchSubwayActivity.kt
@@ -1,13 +1,14 @@
 package com.playtogether_android.app.presentation.ui.onboarding
 
 import android.os.Bundle
+import android.text.Editable
+import android.text.TextWatcher
 import androidx.activity.viewModels
 import com.playtogether_android.app.R
 import com.playtogether_android.app.databinding.ActivitySearchSubwayBinding
 import com.playtogether_android.app.presentation.base.BaseActivity
 import com.playtogether_android.app.presentation.ui.onboarding.adapter.*
 import com.playtogether_android.app.presentation.ui.onboarding.viewmodel.OnBoardingViewModel
-import com.playtogether_android.domain.model.onboarding.SubwayData
 import com.playtogether_android.domain.model.onboarding.SubwayListData
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
@@ -17,30 +18,12 @@ class SearchSubwayActivity :
     BaseActivity<ActivitySearchSubwayBinding>(R.layout.activity_search_subway) {
 
     private val onBoardingViewModel: OnBoardingViewModel by viewModels()
-    // private lateinit var onboardingSubwayAdapter: OnboardingSubwayAdapter
-
-//    private var recipeList = onBoardingViewModel.recipeList
-
-    private lateinit var recipeAdapter: SearchAfterAdapter
-    private lateinit var resultAdapter: SearchResultAdapter
-    private lateinit var ingredientAdapter: SearchInSubwayAdapter
+    private lateinit var subwayAdapter: SubwayAdapter
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        subwayNetwork()
         getList()
 
-    }
-
-
-    private fun subwayNetwork() {
-
-        onBoardingViewModel.getSubwayList()
-        //onboardingSubwayAdapter = OnboardingSubwayAdapter()
-        onBoardingViewModel.subwayList.observe(this) {
-            Timber.d("SubwayList : $it")
-            //onboardingSubwayAdapter.setSubwayList((it) as MutableList<SubwayListData>)
-        }
     }
 
 
@@ -56,37 +39,41 @@ class SearchSubwayActivity :
     private fun addList() {
         onBoardingViewModel.listAddAll.observe(this) {
             if (it == true) {
-                observingWord()
+                //observingWord()
+                editTextWatcher()
             }
         }
     }
 
+    private fun editTextWatcher() = with(binding) {
+        etSubwayOnboardingName.addTextChangedListener(object : TextWatcher {
+            override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
 
-    private fun observingWord() {
-        searchingRecipes("화")
-        onBoardingViewModel.searchingWord.observe(this) {
-            Timber.d("TEST WORD: $it")
+            }
 
-            //           searchingIngredients(it)
-        }
+            override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
 
+            }
 
+            override fun afterTextChanged(p0: Editable?) {
+                val input = binding.etSubwayOnboardingName.text.toString()
+
+                searchingRecipes(input)
+            }
+        })
     }
 
     private fun searchingRecipes(text: String) {
         val tmpList = ArrayList<SubwayListData>()
         val searchSubwayList = onBoardingViewModel.searchSubwayList
-        Timber.d("kkkkkkkk : $searchSubwayList")
-        for (i in 0..searchSubwayList.size - 1) {
+        for (i in 0 until searchSubwayList.size) {
             if (searchSubwayList.get(i).STATION_NM.contains(text)) {
                 tmpList.add(searchSubwayList.get(i))
             }
         }
-//        recipeAdapter.findText = text
-        recipeAdapter = SearchAfterAdapter()
-        binding.rvOnboardingSubway.adapter = recipeAdapter
-        recipeAdapter.setCrewList(tmpList)
+        subwayAdapter = SubwayAdapter()
+        binding.rvOnboardingSubway.adapter = subwayAdapter
+        subwayAdapter.findText = text
+        subwayAdapter.setCrewList(tmpList)
     }
-
-
 }
