+++ app/src/main/java/com/playtogether_android/app/presentation/ui/onboarding/adapter/SubwayAdapter.kt
@@ -1,10 +1,11 @@
 package com.playtogether_android.app.presentation.ui.onboarding.adapter
 
+import android.text.Spannable
+import android.text.SpannableStringBuilder
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.playtogether_android.app.databinding.ItemOnboardingSubwayBinding
-import com.playtogether_android.domain.model.onboarding.CrewListData
 import com.playtogether_android.domain.model.onboarding.SubwayListData
 
 /*
@@ -65,9 +66,10 @@ class SearchAfterAdapter(private val itemClick: (SubwayListData) -> Unit) :
 
  */
 
-class SearchAfterAdapter() :
-    RecyclerView.Adapter<SearchAfterAdapter.OnboardingListViewHolder>() {
+class SubwayAdapter() :
+    RecyclerView.Adapter<SubwayAdapter.OnboardingListViewHolder>() {
     var dataList = mutableListOf<SubwayListData>()
+    var findText = ""
 
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): OnboardingListViewHolder {
@@ -80,7 +82,7 @@ class SearchAfterAdapter() :
     }
 
     override fun onBindViewHolder(holder: OnboardingListViewHolder, position: Int) {
-        holder.onBind(dataList[position])
+        holder.onBind(dataList[position], makeBold(dataList[position].STATION_NM, findText))
     }
 
     override fun getItemCount(): Int = dataList.size
@@ -88,8 +90,9 @@ class SearchAfterAdapter() :
     class OnboardingListViewHolder(
         val binding: ItemOnboardingSubwayBinding
     ) : RecyclerView.ViewHolder(binding.root) {
-        fun onBind(data: SubwayListData) {
+        fun onBind(data: SubwayListData, text: SpannableStringBuilder) {
             binding.apply {
+                tvSubwayName.text = text
                 subway = data
                 executePendingBindings()
             }
@@ -99,5 +102,18 @@ class SearchAfterAdapter() :
     fun setCrewList(data: MutableList<SubwayListData>) {
         this.dataList = data
         notifyDataSetChanged()
+    }
+
+    fun makeBold(fulltext: String, findText: String): SpannableStringBuilder {
+        val str = SpannableStringBuilder(fulltext)
+        val startInt = fulltext.indexOf(findText)
+        val endInt = startInt + findText.length
+        str.setSpan(
+            android.text.style.StyleSpan(android.graphics.Color.RED),
+            startInt,
+            endInt,
+            Spannable.SPAN_INCLUSIVE_EXCLUSIVE
+        )
+        return str
     }
 }
